requirements.txt
Django==5.1.4
djangorestframework==3.14.0
django-filter==24.1
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.1.0
python-dateutil==2.8.2
requests==2.31.0
PyJWT==2.8.1
simplejson==3.19.2
